@@ -31,7 +31,8 @@
   },
   "devDependencies": {
     "dotenv": "^10.0.0",
-    "eslint": "^7.28.0",
+    "eslint": "^7.30.0",
+    "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-node": "^11.1.0",
