@@ -1,5 +1,5 @@
-const { sequelize } = require('./index');
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
+import sequelize from './index';
 
 // This is the junction table between users and campaigns
 const CampaignDetail = sequelize.define('CampaignDetail', {
@@ -8,18 +8,18 @@ const CampaignDetail = sequelize.define('CampaignDetail', {
     type: DataTypes.INTEGER,
     defaultValue: 1,
     validate: {
-      isIn: [[0, 1]]
-    }
+      isIn: [[0, 1]],
+    },
   },
   mechCredRollsRemain: {
     type: DataTypes.INTEGER,
     defaultValue: 1,
     validate: {
-      isIn: [[0, 1]]
-    }
-  }
+      isIn: [[0, 1]],
+    },
+  },
 });
 
 CampaignDetail.sync();
 
-module.exports = CampaignDetail;
+export default CampaignDetail;
