@@ -1,15 +1,15 @@
-const { sequelize } = require('./index');
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
+import sequelize from './index';
 
 const Campaign = sequelize.define('Campaign', {
   name: {
     type: DataTypes.STRING,
-    allowNull: false
+    allowNull: false,
   },
   startDate: DataTypes.DATE,
-  endDate: DataTypes.DATE
+  endDate: DataTypes.DATE,
 });
 
 Campaign.sync();
 
-module.exports = Campaign;
+export default Campaign;
