@@ -1,16 +1,16 @@
-const { sequelize } = require('./index');
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
+import { sequelize } from './index';
 
 const Operation = sequelize.define('Operation', {
   name: {
     type: DataTypes.STRING,
-    allowNull: false
+    allowNull: false,
   },
   objectives: DataTypes.ARRAY(DataTypes.STRING),
   startDate: DataTypes.STRING,
-  endDate: DataTypes.STRING
+  endDate: DataTypes.STRING,
 });
 
 Operation.sync();
 
-module.exports = Operation;
+export default Operation;
