@@ -1,14 +1,14 @@
-const { sequelize } = require('./index');
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
+import { sequelize } from './index';
 
 const Pilot = sequelize.define('Pilot', {
   firstName: {
     type: DataTypes.STRING,
-    allowNull: false
+    allowNull: false,
   },
   lastName: {
     type: DataTypes.STRING,
-    allowNull: false
+    allowNull: false,
   },
   birth: DataTypes.DATE,
   death: DataTypes.DATE,
@@ -18,29 +18,29 @@ const Pilot = sequelize.define('Pilot', {
   plusMovement: DataTypes.INTEGER,
   accuracyShort: {
     type: DataTypes.INTEGER,
-    defaultValue: 2
+    defaultValue: 2,
   },
   accuracyMed: {
     type: DataTypes.INTEGER,
-    defaultValue: 4
+    defaultValue: 4,
   },
   accuracyLong: {
     type: DataTypes.INTEGER,
-    defaultValue: 6
+    defaultValue: 6,
   },
   plusDamageShort: DataTypes.INTEGER,
   plusDamageMed: DataTypes.INTEGER,
   plusDamageLong: DataTypes.INTEGER,
   baseTargetNum: {
     type: DataTypes.INTEGER,
-    defaultValue: 4
+    defaultValue: 4,
   },
   totalMovementModifier: {
     type: DataTypes.INTEGER,
-    defaultValue: 1
-  }
+    defaultValue: 1,
+  },
 });
 
 Pilot.sync();
 
-module.exports = Pilot;
+export default Pilot;
